+++ piwigo/Image/Parameters/Delegates/EditImageParamsViewController+DatePicker.swift
@@ -1,0 +1,39 @@
+//
+//  EditImageParamsViewController+DatePicker.swift
+//  piwigo
+//
+//  Created by Eddy Lelièvre-Berna on 31/12/2023.
+//  Copyright © 2023 Piwigo.org. All rights reserved.
+//
+
+// MARK: -  EditImageDatePickerDelegate Methods
+extension EditImageParamsViewController: EditImageDatePickerDelegate
+{
+    func didSelectDate(withPicker date: Date) {
+        // Apply new date
+        shouldUpdateDateCreated = true
+        commonDateCreated = date
+
+        // Update cell
+        let indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
+        editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
+    }
+
+    func didUnsetImageCreationDate() {
+        commonDateCreated = .distantPast
+        shouldUpdateDateCreated = true
+
+        // Close date picker
+        var indexPath = IndexPath(row: EditImageParamsOrder.datePicker.rawValue, section: 0)
+        if hasDatePicker {
+            hasDatePicker = false
+            editImageParamsTableView.beginUpdates()
+            editImageParamsTableView.deleteRows(at: [indexPath], with: .fade)
+            editImageParamsTableView.endUpdates()
+        }
+
+        // Update creation date cell
+        indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
+        editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
+    }
+}
