+++ piwigo/Image/Parameters/EditImageParamsViewController.swift
@@ -24,32 +24,32 @@ class EditImageParamsViewController: UIViewController
     var hasTagCreationRights = false
     weak var delegate: EditImageParamsDelegate?
     
-    @IBOutlet private weak var editImageParamsTableView: UITableView!
+    @IBOutlet weak var editImageParamsTableView: UITableView!
     private var hudViewController: UIViewController?
     private let kEditImageParamsViewWidth: CGFloat = 512.0
 
-    private var shouldUpdateTitle = false
-    private var commonTitle = NSAttributedString()
+    var shouldUpdateTitle = false
+    var commonTitle = NSAttributedString()
     
-    private var shouldUpdateAuthor = false
-    private var commonAuthor = ""
+    var shouldUpdateAuthor = false
+    var commonAuthor = ""
     
-    private var hasDatePicker = false
-    private var shouldUpdateDateCreated = false
-    private var commonDateCreated = Date.distantPast
-    private var oldCreationDate = Date()
+    var hasDatePicker = false
+    var shouldUpdateDateCreated = false
+    var commonDateCreated = Date.distantPast
+    var oldCreationDate = Date()
     private var timeOffset = TimeInterval.zero
 
-    private var shouldUpdatePrivacyLevel = false
-    private var commonPrivacyLevel = pwgPrivacy.everybody.rawValue
+    var shouldUpdatePrivacyLevel = false
+    var commonPrivacyLevel = pwgPrivacy.everybody.rawValue
     
-    private var shouldUpdateTags = false
-    private var commonTags = Set<Tag>()
-    private var addedTags = Set<Tag>()
-    private var removedTags = Set<Tag>()
+    var shouldUpdateTags = false
+    var commonTags = Set<Tag>()
+    var addedTags = Set<Tag>()
+    var removedTags = Set<Tag>()
     
-    private var shouldUpdateComment = false
-    private var commonComment = NSAttributedString()
+    var shouldUpdateComment = false
+    var commonComment = NSAttributedString()
     
     enum EditImageParamsOrder : Int {
         case thumbnails
@@ -485,559 +485,6 @@ class EditImageParamsViewController: UIViewController
             completion()
         } failure: { error in
             failure(error)
-        }
-    }
-}
-
-
-// MARK: - UITableViewDataSource Methods
-extension EditImageParamsViewController: UITableViewDataSource
-{
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        var nberOfRows = EditImageParamsOrder.count.rawValue
-        nberOfRows -= hasDatePicker ? 0 : 1
-        nberOfRows -= user.hasAdminRights ? 0 : 1
-        return nberOfRows
-    }
-
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        var height: CGFloat = 44.0
-        var row = indexPath.row
-        row += (!hasDatePicker && (row > EditImageParamsOrder.date.rawValue)) ? 1 : 0
-        row += (!user.hasAdminRights && (row > EditImageParamsOrder.datePicker.rawValue)) ? 1 : 0
-        switch EditImageParamsOrder(rawValue: row) {
-            case .thumbnails:
-                height = 188.0
-            case .datePicker:
-                if images.count > 1 {
-                    // Time interval picker
-                    height = 258.0
-                } else {
-                    // Date picker
-                    height = 304.0
-                }
-            case .privacy, .tags:
-                height = 73.0
-            case .desc:
-                height = 428.0
-            default:
-                break
-        }
-
-        return height
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        var tableViewCell = UITableViewCell()
-
-        var row = indexPath.row
-        row += (!hasDatePicker && (row > EditImageParamsOrder.date.rawValue)) ? 1 : 0
-        row += (!user.hasAdminRights && (row > EditImageParamsOrder.datePicker.rawValue)) ? 1 : 0
-        switch EditImageParamsOrder(rawValue: row) {
-        case .thumbnails:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "EditImageThumbTableViewCell", for: indexPath) as? EditImageThumbTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageThumbTableViewCell!")
-                return EditImageThumbTableViewCell()
-            }
-            cell.config(withImages: images)
-            cell.delegate = self
-            tableViewCell = cell
-            
-        case .imageName:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "title", for: indexPath) as? EditImageTextFieldTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageTextFieldTableViewCell!")
-                return EditImageTextFieldTableViewCell()
-            }
-            let wholeRange = NSRange(location: 0, length: commonTitle.string.count)
-            let style = NSMutableParagraphStyle()
-            style.alignment = NSTextAlignment.right
-            let attributes = [
-                NSAttributedString.Key.font: UIFont.systemFont(ofSize: 17),
-                NSAttributedString.Key.paragraphStyle: style
-            ]
-            let detail = NSMutableAttributedString(attributedString: commonTitle)
-            detail.addAttributes(attributes, range: wholeRange)
-            cell.config(withLabel: NSAttributedString(string: NSLocalizedString("editImageDetails_title", comment: "Title")), placeHolder: NSLocalizedString("editImageDetails_titlePlaceholder", comment: "Title"), andImageDetail: detail)
-            if shouldUpdateTitle {
-                cell.cellTextField.textColor = .piwigoColorOrange()
-            }
-            cell.cellTextField.tag = row
-            cell.cellTextField.delegate = self
-            tableViewCell = cell
-            
-        case .author:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "author", for: indexPath) as? EditImageTextFieldTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageTextFieldTableViewCell!")
-                return EditImageTextFieldTableViewCell()
-            }
-            cell.config(withLabel: NSAttributedString(string: NSLocalizedString("editImageDetails_author", comment: "Author")), placeHolder: NSLocalizedString("settings_defaultAuthorPlaceholder", comment: "Author Name"), andImageDetail: NSAttributedString(string: commonAuthor))
-            if shouldUpdateAuthor {
-                cell.cellTextField.textColor = .piwigoColorOrange()
-            }
-            cell.cellTextField.tag = row
-            cell.cellTextField.delegate = self
-            tableViewCell = cell
-            
-        case .date:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "dateCreation", for: indexPath) as? EditImageTextFieldTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageTextFieldTableViewCell!")
-                return EditImageTextFieldTableViewCell()
-            }
-            cell.config(withLabel: NSAttributedString(string: NSLocalizedString("editImageDetails_dateCreation", comment: "Creation Date")), placeHolder: "", andImageDetail: NSAttributedString(string: getStringFrom(commonDateCreated)))
-            if shouldUpdateDateCreated {
-                cell.cellTextField.textColor = .piwigoColorOrange()
-            }
-            cell.cellTextField.tag = row
-            cell.cellTextField.delegate = self
-            tableViewCell = cell
-            
-        case .datePicker:
-            // Which picker?
-            if images.count > 1 {
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: "ShiftPickerTableCell", for: indexPath) as? EditImageShiftPickerTableViewCell else {
-                    print("Error: tableView.dequeueReusableCell does not return a EditImageShiftPickerTableViewCell!")
-                    return EditImageShiftPickerTableViewCell()
-                }
-                cell.config(withDate: commonDateCreated, animated: false)
-                cell.delegate = self
-                tableViewCell = cell
-                
-            } else {
-                guard let cell = tableView.dequeueReusableCell(withIdentifier: "DatePickerTableCell", for: indexPath) as? EditImageDatePickerTableViewCell else {
-                    print("Error: tableView.dequeueReusableCell does not return a EditImageDatePickerTableViewCell!")
-                    return EditImageDatePickerTableViewCell()
-                }
-                cell.config(withDate: commonDateCreated, animated: false)
-                cell.setDatePickerButtons()
-                cell.delegate = self
-                tableViewCell = cell
-            }
-            
-        case .privacy:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "privacy", for: indexPath) as? EditImagePrivacyTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImagePrivacyTableViewCell!")
-                return EditImagePrivacyTableViewCell()
-            }
-            cell.setLeftLabel(withText: NSLocalizedString("editImageDetails_privacyLevel", comment: "Who can see this photo?"))
-            cell.setPrivacyLevel(with: pwgPrivacy(rawValue: commonPrivacyLevel) ?? .everybody,
-                                 inColor: shouldUpdatePrivacyLevel ? .piwigoColorOrange() : .piwigoColorRightLabel())
-            tableViewCell = cell
-            
-        case .tags:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "tags", for: indexPath) as? EditImageTagsTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageTagsTableViewCell!")
-                return EditImageTagsTableViewCell()
-            }
-            cell.config(withList: commonTags,
-                        inColor: shouldUpdateTags ? UIColor.piwigoColorOrange() : UIColor.piwigoColorRightLabel())
-            tableViewCell = cell
-            
-        case .desc:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "description", for: indexPath) as? EditImageTextViewTableViewCell else {
-                print("Error: tableView.dequeueReusableCell does not return a EditImageTextViewTableViewCell!")
-                return EditImageTextViewTableViewCell()
-            }
-            let wholeRange = NSRange(location: 0, length: commonComment.string.count)
-            let style = NSMutableParagraphStyle()
-            style.alignment = NSTextAlignment.left
-            let attributes = [
-                NSAttributedString.Key.font: UIFont.systemFont(ofSize: 17),
-                NSAttributedString.Key.paragraphStyle: style
-            ]
-            let detail = NSMutableAttributedString(attributedString: commonComment)
-            detail.addAttributes(attributes, range: wholeRange)
-            cell.config(withText: detail,
-                        inColor: shouldUpdateTags ? .piwigoColorOrange() : .piwigoColorRightLabel())
-            cell.textView.delegate = self
-            tableViewCell = cell
-            
-        default:
-            break
-        }
-
-        tableViewCell.backgroundColor = .piwigoColorCellBackground()
-        tableViewCell.tintColor = .piwigoColorOrange()
-        return tableViewCell
-    }
-
-    private func getStringFrom(_ date: Date?) -> String {
-        var dateStr = ""
-        var timeStr = ""
-        if let date = date {
-            if view.bounds.size.width > 430 {
-                // i.e. larger than iPhone 14 Pro Max screen width
-                dateStr = DateFormatter.localizedString(from: date, dateStyle: .long, timeStyle: .none)
-                timeStr = DateFormatter.localizedString(from: date, dateStyle: .none, timeStyle: .medium)
-            } else if view.bounds.size.width > 320 {
-                // i.e. larger than iPhone 5 screen width
-                dateStr = DateFormatter.localizedString(from: date, dateStyle: .medium, timeStyle: .none)
-                timeStr = DateFormatter.localizedString(from: date, dateStyle: .none, timeStyle: .medium)
-            } else {
-                dateStr = DateFormatter.localizedString(from: date, dateStyle: .short, timeStyle: .none)
-                timeStr = DateFormatter.localizedString(from: date, dateStyle: .none, timeStyle: .medium)
-            }
-        }
-        return "\(dateStr) - \(timeStr)"
-    }
-}
-    
- 
-// MARK: - UITableViewDelegate Methods
-extension EditImageParamsViewController: UITableViewDelegate
-{
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 0.0 // To hide the section header
-    }
-
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return 0.0 // To hide the section footer
-    }
-
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        var row = indexPath.row
-        row += (!hasDatePicker && (row > EditImageParamsOrder.date.rawValue)) ? 1 : 0
-        row += (!user.hasAdminRights && (row > EditImageParamsOrder.datePicker.rawValue)) ? 1 : 0
-        switch EditImageParamsOrder(rawValue: row) {
-        case .privacy:
-            // Deselect row
-            tableView.deselectRow(at: indexPath, animated: true)
-
-            // Dismiss the keyboard
-            view.endEditing(true)
-
-            // Create view controller
-            let privacySB = UIStoryboard(name: "SelectPrivacyViewController", bundle: nil)
-            guard let privacyVC = privacySB.instantiateViewController(withIdentifier: "SelectPrivacyViewController") as? SelectPrivacyViewController else { return }
-            privacyVC.delegate = self
-            privacyVC.privacy = pwgPrivacy(rawValue: commonPrivacyLevel) ?? .everybody
-            navigationController?.pushViewController(privacyVC, animated: true)
-            
-        case .tags:
-            // Deselect row
-            tableView.deselectRow(at: indexPath, animated: true)
-
-            // Dismiss the keyboard
-            view.endEditing(true)
-
-            // Create view controller
-            let tagsSB = UIStoryboard(name: "TagsViewController", bundle: nil)
-            guard let tagsVC = tagsSB.instantiateViewController(withIdentifier: "TagsViewController") as? TagsViewController else { return }
-            tagsVC.delegate = self
-            tagsVC.user = user
-            let tagList: [Int32] = commonTags.compactMap { Int32($0.tagId) }
-            tagsVC.setPreselectedTagIds(Set(tagList))
-            navigationController?.pushViewController(tagsVC, animated: true)
-            
-        default:
-            break
-        }
-    }
-
-    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
-        var result: Bool
-        var row = indexPath.row
-        row += (!hasDatePicker && (row > EditImageParamsOrder.date.rawValue)) ? 1 : 0
-        row += (!user.hasAdminRights && (row > EditImageParamsOrder.datePicker.rawValue)) ? 1 : 0
-        switch EditImageParamsOrder(rawValue: row) {
-            case .imageName, .author, .date, .datePicker, .desc:
-                result = false
-            default:
-                result = true
-        }
-
-        return result
-    }
-}
-    
- 
-// MARK: - UITextFieldDelegate Methods
-extension EditImageParamsViewController: UITextFieldDelegate
-{
-    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
-        if textField.tag == EditImageParamsOrder.date.rawValue {
-            // The common date can be distant past (i.e. unset)
-            if commonDateCreated == .distantPast {
-                // Define date as today
-                commonDateCreated = Date()
-                shouldUpdateDateCreated = true
-
-                // Update creation date
-                let indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
-                editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
-            }
-
-            // Show date or hide picker
-            let indexPath = IndexPath(row: EditImageParamsOrder.datePicker.rawValue, section: 0)
-            if hasDatePicker {
-                // Found a picker, so remove it
-                hasDatePicker = false
-                editImageParamsTableView.beginUpdates()
-                editImageParamsTableView.deleteRows(at: [indexPath], with: .fade)
-                editImageParamsTableView.endUpdates()
-            } else {
-                // Didn't find a picker, so we should insert it
-                hasDatePicker = true
-                editImageParamsTableView.beginUpdates()
-                editImageParamsTableView.insertRows(at: [indexPath], with: .fade)
-                editImageParamsTableView.endUpdates()
-            }
-
-            // Prevent keyboard from opening
-            return false
-        }
-
-        return true
-    }
-
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        switch EditImageParamsOrder(rawValue: textField.tag) {
-            case .imageName:
-                // Title
-                shouldUpdateTitle = true
-                textField.textColor = .piwigoColorOrange()
-            case .author:
-                // Author
-                shouldUpdateAuthor = true
-                textField.textColor = .piwigoColorOrange()
-            default:
-                break
-        }
-    }
-
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange,
-                   replacementString string: String) -> Bool {
-        guard let finalString = (textField.text as NSString?)?.replacingCharacters(in: range, with: string)
-            else { return false }
-        switch EditImageParamsOrder(rawValue: textField.tag) {
-            case .imageName:
-                // Title
-                commonTitle = finalString.htmlToAttributedString
-            case .author:
-                // Author
-                commonAuthor = finalString
-            default:
-                break
-        }
-        return true
-    }
-
-    func textFieldShouldClear(_ textField: UITextField) -> Bool {
-        switch EditImageParamsOrder(rawValue: textField.tag) {
-            case .imageName:
-                // Title
-            commonTitle = "".htmlToAttributedString
-            case .author:
-                // Author
-                commonAuthor = ""
-            default:
-                break
-        }
-        return true
-    }
-
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        editImageParamsTableView.endEditing(true)
-        return true
-    }
-
-    func textFieldDidEndEditing(_ textField: UITextField) {
-        switch EditImageParamsOrder(rawValue: textField.tag) {
-            case .imageName:
-                // Title
-            commonTitle = (textField.text ?? "").htmlToAttributedString
-            case .author:
-                // Author
-                commonAuthor = textField.text ?? ""
-            default:
-                break
-        }
-    }
-}
-
-
-// MARK: - UITextViewDelegate Methods
-extension EditImageParamsViewController: UITextViewDelegate
-{
-    func textViewDidBeginEditing(_ textView: UITextView) {
-        shouldUpdateComment = true
-        textView.textColor = .piwigoColorOrange()
-    }
-
-    func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText text: String) -> Bool {
-        let finalString = (textView.text as NSString).replacingCharacters(in: range, with: text)
-        commonComment = finalString.htmlToAttributedString
-        return true
-    }
-
-    func textViewShouldEndEditing(_ textView: UITextView) -> Bool {
-        editImageParamsTableView.endEditing(true)
-        return true
-    }
-
-    func textViewDidEndEditing(_ textView: UITextView) {
-        commonComment = textView.text.htmlToAttributedString
-    }
-}
-
-
-// MARK: - EditImageThumbnailCellDelegate Methods
-extension EditImageParamsViewController: EditImageThumbnailCellDelegate
-{
-    func didDeselectImage(withId imageId: Int64) {
-        // Hide picker if needed
-        let indexPath = IndexPath(row: EditImageParamsOrder.datePicker.rawValue, section: 0)
-        if hasDatePicker {
-            // Found a picker, so remove it
-            hasDatePicker = false
-            editImageParamsTableView.beginUpdates()
-            editImageParamsTableView.deleteRows(at: [indexPath], with: .fade)
-            editImageParamsTableView.endUpdates()
-        }
-
-        // Update data source
-        let timeInterval = commonDateCreated.timeIntervalSince(oldCreationDate)
-        images.removeAll(where: {$0.pwgID == imageId})
-
-        // Update common creation date if needed
-        oldCreationDate = Date(timeIntervalSinceReferenceDate: images[0].dateCreated)
-        commonDateCreated = oldCreationDate.addingTimeInterval(timeInterval)
-
-        // Refresh table
-        editImageParamsTableView.reloadData()
-
-        // Deselect image in album view
-        delegate?.didDeselectImage(withId: imageId)
-    }
-
-    func didRenameFileOfImage(_ imageData: Image) {
-        // Update data source
-        do {
-            try mainContext.save()
-        } catch let error as NSError {
-            print("Could not fetch \(error), \(error.userInfo)")
-        }
-
-        // Update parent image view
-        delegate?.didChangeImageParameters(imageData)
-    }
-}
-
-
-// MARK: -  EditImageDatePickerDelegate Methods
-extension EditImageParamsViewController: EditImageDatePickerDelegate
-{
-    func didSelectDate(withPicker date: Date) {
-        // Apply new date
-        shouldUpdateDateCreated = true
-        commonDateCreated = date
-
-        // Update cell
-        let indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
-        editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
-    }
-
-    func didUnsetImageCreationDate() {
-        commonDateCreated = .distantPast
-        shouldUpdateDateCreated = true
-
-        // Close date picker
-        var indexPath = IndexPath(row: EditImageParamsOrder.datePicker.rawValue, section: 0)
-        if hasDatePicker {
-            hasDatePicker = false
-            editImageParamsTableView.beginUpdates()
-            editImageParamsTableView.deleteRows(at: [indexPath], with: .fade)
-            editImageParamsTableView.endUpdates()
-        }
-
-        // Update creation date cell
-        indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
-        editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
-    }
-}
-
-
-// MARK: -  EditImageShiftPickerDelegate Methods
-extension EditImageParamsViewController: EditImageShiftPickerDelegate
-{
-    func didSelectDate(withShiftPicker date: Date) {
-        // Apply new date
-        shouldUpdateDateCreated = true
-        commonDateCreated = date
-
-        // Update cell
-        let indexPath = IndexPath(row: EditImageParamsOrder.date.rawValue, section: 0)
-        editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
-    }
-}
-
-
-// MARK: - SelectPrivacyObjcDelegate Methods
-extension EditImageParamsViewController: SelectPrivacyDelegate
-{
-    func didSelectPrivacyLevel(_ privacyLevel: pwgPrivacy) {
-        // Check if the user decided to leave the Edit mode
-        if !(navigationController?.visibleViewController is EditImageParamsViewController) {
-            // Return updated parameters
-            if delegate?.responds(to: #selector(EditImageParamsDelegate.didFinishEditingParameters)) ?? false {
-                delegate?.didFinishEditingParameters()
-            }
-            return
-        }
-
-        // Update image parameter?
-        if privacyLevel.rawValue != commonPrivacyLevel {
-            // Remember to update image info
-            shouldUpdatePrivacyLevel = true
-            commonPrivacyLevel = privacyLevel.rawValue
-
-            // Refresh table row
-            let row = EditImageParamsOrder.privacy.rawValue - (hasDatePicker == false ? 1 : 0)
-            let indexPath = IndexPath(row: row, section: 0)
-            editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
-        }
-    }
-}
-
-
-// MARK: - TagsViewControllerObjcDelegate Methods
-extension EditImageParamsViewController: TagsViewControllerDelegate
-{
-    func didSelectTags(_ selectedTags: Set<Tag>) {
-        // Check if the user decided to leave the Edit mode
-        if !(navigationController?.visibleViewController is EditImageParamsViewController) {
-            // Return updated parameters
-            delegate?.didFinishEditingParameters()
-            return
-        }
-
-        // Build list of added tags
-        addedTags = []
-        for tag in selectedTags {
-            if commonTags.contains(where: { $0.tagId == tag.tagId }) == false {
-                addedTags.insert(tag)
-            }
-        }
-
-        // Build list of removed tags
-        removedTags = []
-        for tag in commonTags {
-            if !selectedTags.contains(where: { $0.tagId == tag.tagId }) {
-                removedTags.insert(tag)
-            }
-        }
-
-        // Do we need to update images?
-        if (addedTags.isEmpty == false) || (removedTags.isEmpty == false) {
-            // Update common tag list and remember to update image info
-            shouldUpdateTags = true
-            commonTags = Set(selectedTags)
-
-            // Refresh table row
-            var row: Int = EditImageParamsOrder.tags.rawValue
-            row -= !hasDatePicker ? 1 : 0
-            row -= !user.hasAdminRights ? 1 : 0
-            let indexPath = IndexPath(row: row, section: 0)
-            editImageParamsTableView.reloadRows(at: [indexPath], with: .automatic)
         }
     }
 }
